feat: extend circuit breaker with exception filtering and anti-flapping, enrich cache context with hot-key controls, and migrate to long counters with batch preload
The circuit breaker previously had no way to distinguish between system failures and business exceptions, causing spurious trips on client errors. Its half-open recovery relied on a single probe, which could flip the breaker open again due to a transient glitch. On the annotation side, the per-method context only provided TTL overrides and a broadcast skip flag, leaving hot-key specific TTLs and detection suppression impossible to configure directly on individual `@Cacheable` methods. Preload annotation used `int` increments, limiting the maximum inflation to ~2 billion, and required separate calls per key, adding overhead for bulk pre-warming.

To address these issues, the circuit breaker now accepts `exclude-exceptions` and `include-exceptions` lists — when a thrown exception or its cause matches the exclude list, it is treated as a success and does not count toward the failure rate; when the include list is non-empty, only those exceptions trip the breaker. The half-open state now requires `consecutive-success-threshold` consecutive successes before closing, preventing the breaker from oscillating due to a single lucky probe. Internally, the window bucket counters were replaced with a padding‑aware `long[]` array accessed via `VarHandle` to reduce false sharing and improve throughput under contention. The `ZetaCacheContext` record was expanded with fields for hot‑key hard/soft TTL overrides and a skip‑detection flag, making it possible to fine‑tune hot‑key behaviour at the method level without global defaults. New annotations `@HotTTL`, `@CacheCondition`, `@SkipBroadcast`, and `@SkipDetection` were introduced, while the old `@Broadcast` annotation was removed. The `@CacheTTL` annotation gained SpEL support via `hardTtlSpEl` / `softTtlSpEl`, enabling dynamic TTLs based on method parameters. Preload’s `count()` was widened to `long` and a batch `notifyLocalDetectorDirect(Map)` method was added to register multiple keys in a single call; the underlying `TopK.addDirect` signature changed from `int` to `long` accordingly.

After these changes, the circuit breaker correctly ignores client‑side exceptions (e.g. `IllegalArgumentException`) and resists flapping in wide networks or under variable load. The cache context exposes hot‑key specific TTLs and detection bypass, allowing more precise caching strategies per endpoint. Preload keys can now exceed 2 billion counts and are registered in bulk, reducing overhead. The `VarHandle`‑based counters eliminate `LongAdder` boxing and improve cache line utilisation. All existing unit tests were adapted and pass, while new tests verify exception filtering, consecutive‑success closing, and the extended context fields.

+++ common/src/main/java/io/github/hyshmily/zeta/Zeta.java
@@ -1606,7 +1606,7 @@ public class Zeta implements DisposableBean {
    * @param cacheKey the key to record
    * @param count    the number of accesses to add
    */
-  public void notifyLocalDetectorDirect(String cacheKey, int count) {
+  public void notifyLocalDetectorDirect(String cacheKey, long count) {
     requireAppDetector("notifyLocalDetectorDirect");
     appHotKeyDetector.addDirect(cacheKey, count);
   }
